Fix constructor pixel_snap for all drawables

+++ src/fe_listbox.cpp
@@ -31,7 +31,7 @@
 
 const std::string DEFAULT_FORMAT_STRING = "[Title]";
 
-FeListBox::FeListBox( FePresentableParent &p, int x, int y, int w, int h )
+FeListBox::FeListBox( FePresentableParent &p, float x, float y, float w, float h )
 	: FeBasePresentable( p ),
 	m_selColour( sf::Color::Yellow ),
 	m_selBg( sf::Color::Blue ),
