+++ src/fe_present.cpp
@@ -741,7 +741,7 @@ FeImage *FePresent::add_clone( FeImage *o,
 	return new_image;
 }
 
-FeText *FePresent::add_text( const std::string &n, int x, int y, int w, int h,
+FeText *FePresent::add_text( const std::string &n, float x, float y, float w, float h,
 			FePresentableParent &p )
 {
 	FeText *new_text = new FeText( p, n, x, y, w, h );
@@ -759,7 +759,7 @@ FeText *FePresent::add_text( const std::string &n, int x, int y, int w, int h,
 	return new_text;
 }
 
-FeListBox *FePresent::add_listbox( int x, int y, int w, int h,
+FeListBox *FePresent::add_listbox( float x, float y, float w, float h,
 		FePresentableParent &p )
 {
 	FeListBox *new_lb = new FeListBox( p, x, y, w, h );
