+++ src/fe_present.hpp
@@ -203,8 +203,8 @@ protected:
 
 	FeImage *add_image(bool a, const std::string &n, float x, float y, float w, float h, FePresentableParent &p);
 	FeImage *add_clone(FeImage *, FePresentableParent &p);
-	FeText *add_text(const std::string &n, int x, int y, int w, int h, FePresentableParent &p);
-	FeListBox *add_listbox(int x, int y, int w, int h, FePresentableParent &p);
+	FeText *add_text(const std::string &n, float x, float y, float w, float h, FePresentableParent &p);
+	FeListBox *add_listbox(float x, float y, float w, float h, FePresentableParent &p);
 	FeRectangle *add_rectangle(float x, float y, float w, float h, FePresentableParent &p);
 	FeImage *add_surface(float x, float y, float w, float h, int texture_width, int texture_height, FePresentableParent &p);
 	FeSound *add_sound(const char *n);
