+++ src/fe_presentable.cpp
@@ -758,7 +758,7 @@ FeImage *FePresentableParent::add_clone(FeImage *i )
 	return NULL;
 }
 
-FeText *FePresentableParent::add_text(const char *t, int x, int y, int w, int h)
+FeText *FePresentableParent::add_text(const char *t, float x, float y, float w, float h)
 {
 	FePresent *fep = FePresent::script_get_fep();
 
@@ -768,7 +768,7 @@ FeText *FePresentableParent::add_text(const char *t, int x, int y, int w, int h)
 	return NULL;
 }
 
-FeListBox *FePresentableParent::add_listbox(int x, int y, int w, int h)
+FeListBox *FePresentableParent::add_listbox(float x, float y, float w, float h)
 {
 	FePresent *fep = FePresent::script_get_fep();
