+++ src/fe_vm.hpp
@@ -240,8 +240,8 @@ public:
 	static FeImage *cb_add_artwork(const char *, float, float);
 	static FeImage *cb_add_artwork(const char *);
 	static FeImage *cb_add_clone(FeImage *);
-	static FeText *cb_add_text(const char *,int, int, int, int);
-	static FeListBox *cb_add_listbox(int, int, int, int);
+	static FeText *cb_add_text(const char *,float, float, float, float);
+	static FeListBox *cb_add_listbox(float, float, float, float);
 	static FeRectangle *cb_add_rectangle(float, float, float, float);
 	static FeImage *cb_add_surface(float, float, float, float);
 	static FeImage *cb_add_surface(float, float, float, float, int, int);
